EPD-612: PR feedback
